requirements.txt
fastapi
websockets
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
aiomysql
PyMySQL
redis>=4.2.0rc1
pytz
httpx
tenacity>=8.2.0
openai
groq
mutagen
python-multipart
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
sqlalchemy
email-validator
emails
watchdog
pyyaml
importlib-metadata>=4.6
